requirements.txt
Django
celery
django-colorfield
django-admin-sortable2
grpcio-tools
mypy-protobuf
docker